Moved AddManageLibrariesGroup to com.google.cloud.tools.eclipse.appengine.ui.
