+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/META-INF/MANIFEST.MF
@@ -22,17 +22,10 @@ Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
  com.google.cloud.tools.eclipse.appengine.libraries,
  com.google.cloud.tools.eclipse.appengine.libraries.model,
  com.google.cloud.tools.eclipse.preferences,
- com.google.cloud.tools.eclipse.ui.util.databinding,
  com.google.cloud.tools.eclipse.usagetracker,
  com.google.cloud.tools.eclipse.util.templates.appengine,
  com.google.common.annotations;version="15.0.0",
  com.google.common.base;version="15.0.0",
- org.eclipse.core.databinding,
- org.eclipse.core.databinding.beans,
- org.eclipse.core.databinding.conversion,
- org.eclipse.core.databinding.observable,
- org.eclipse.core.databinding.observable.value,
- org.eclipse.core.databinding.property.value,
  org.eclipse.core.runtime;version="3.5.0",
  org.eclipse.core.runtime.jobs,
  org.eclipse.core.runtime.preferences;version="3.3.0",
