+++ plugins/com.google.cloud.tools.eclipse.appengine.ui/src/com/google/cloud/tools/eclipse/appengine/ui/messages.properties
@@ -1,0 +1,1 @@
+AppEngineStandardWizardPage_librariesGroupLabel=Libraries to add to build path
